Issue #8738: bufferSubData and texImage2D argument sanity checks.

Fixes https://github.com/servo/servo/issues/8738

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/servo/servo/8948)
<!-- Reviewable:end -->

+++ components/canvas_traits/lib.rs
@@ -134,7 +134,7 @@ pub enum CanvasWebGLMsg {
     BlendFuncSeparate(u32, u32, u32, u32),
     AttachShader(u32, u32),
     BindAttribLocation(u32, u32, String),
-    BufferData(u32, Vec<f32>, u32),
+    BufferData(u32, Vec<u8>, u32),
     BufferSubData(u32, isize, Vec<u8>),
     Clear(u32),
     ClearColor(f32, f32, f32, f32),
